backend/requirements.txt
# ===================================
# PlanSight Backend Dependencies
# Merged: Person B (Backend) + Person D (DevOps/Testing)
# ===================================

# === Core Framework (Person B - newer versions) ===
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# === Data Processing ===
numpy==2.1.3

# === Environment & Config ===
python-dotenv==1.0.1

# === LLM Integration ===
google-generativeai==0.8.3

# === HTTP Client (for testing) ===
httpx==0.27.2

# === Person D additions: Load Testing & Development ===
aiohttp==3.9.1  # Async load testing
pytest==7.4.4  # Unit testing
pytest-asyncio==0.23.3  # Async test support

# === Optional: Future Features ===
# Uncomment if implementing /calibrate endpoint
# scikit-learn==1.4.0
# pandas==2.1.4

# === Optional: MongoDB Atlas Integration ===
# Uncomment if using MongoDB for calibration data
# pymongo==4.6.1
# motor==3.3.2
